backend/requirements.txt
fastapi
uvicorn
feedparser
newspaper3k
google-genai
python-dotenv
requests
beautifulsoup4
